Abstract `ClickHouseSink` over the row data type
This commit makes the `ClickHouseSink` generic over the type of data it
inserts into ClickHouse and additionally fixes the issue with hard-coded
table names in DDL queries.

The patch introduces new typeclass `SinkRow` which represents types that
the `ClickHouseSink` can work with. This trait, in addition to requiring
the type to be able to turn itself into `CompactJsonRow`, specifies two
schema-related associated functions: `table_ddl` and `insert_ddl`.

The main benefits of this change (in addition to generality) is better
future testability of the sink. It also adds more safety because now
it's more likely that the DDL matches the data structure.