Add support for blocks of data for insert
This commit reworks the way how buffering is done inside the
`ClickHouseSink`.

It introduces a `InsertQueue` composed of multiple `InsertBlock`s. Each
block has a `MAX_BLOCK_SIZE` and essentially works as the did before
this patch. The queue basically partitions large accumulated datasets of
uncommitted data from Kafka.

The other reason for this change is that ClickHouse itself imposes
insert block size (in the number of rows) and excessively large insert
requests will result in _HTTP 413 (Payload Too Large)_.

Although this `InsertQueue` method helps, it still can hit the 413 if
`MAX_BLOCK_SIZE` is simply taken from the ClickHouse docs. This is
because the underlying client does not (or allow to) do compression.
Therefore the `MAX_BLOCK_SIZE` is set to quite a small value.