Generalize `KafkaSource` over the type of data
This commit abstracts `KafkaConsumer` and consequently `KafkaSource`
over the type of data it parses from a Kafka message and anonymizes.

There is a new trait `Anonymize` representing types that the pipeline
accepts and transformes. `HttpLog` is currently the only instance of
this typeclass.

Additionally, this commit bumps the crate version.